# Zombies recover faster from crit. They heal Piercing damage