--- a/Content.Shared/Zombies/ZombieComponent.cs
+++ b/Content.Shared/Zombies/ZombieComponent.cs
@@ -111,5 +111,13 @@ namespace Content.Shared.Zombies
                 { "Shock", -0.2 },
             }
         };
+
+        [DataField("forceDeathDamage")] public DamageSpecifier ForceDeathDamage = new()
+        {
+            DamageDict = new ()
+            {
+                { "Blunt", 100.0 },
+            }
+        };
     }
 }