--- a/Content.Shared/Zombies/ZombieComponent.cs
+++ b/Content.Shared/Zombies/ZombieComponent.cs
@@ -124,13 +124,5 @@ namespace Content.Shared.Zombies
                 { "Shock", -0.2 },
             }
         };
-
-        [DataField("forceDeathDamage")] public DamageSpecifier ForceDeathDamage = new()
-        {
-            DamageDict = new ()
-            {
-                { "Blunt", 100.0 },
-            }
-        };
     }
 }