--- a/Resources/Locale/en-US/game-ticking/game-presets/preset-zombies.ftl
+++ b/Resources/Locale/en-US/game-ticking/game-presets/preset-zombies.ftl
@@ -5,6 +5,7 @@ zombie-not-enough-ready-players = Not enough players readied up for the game! Th
 zombie-no-one-ready = No players readied up! Can't start Zombies.
 
 zombie-patientzero-role-greeting = You are patient 0. Hide your infection, get supplies, and be prepared to turn once you die.
+zombie-healing = You feel a stirring in your flesh
 
 zombie-alone = You feel entirely alone.