--- a/Resources/Locale/en-US/zombies/zombie.ftl
+++ b/Resources/Locale/en-US/zombies/zombie.ftl
@@ -5,3 +5,5 @@ zombie-generic = zombie
 zombie-name-prefix = Zombified {$target}
 zombie-role-desc =  A malevolent creature of the dead.
 zombie-role-rules = You are an antagonist. Search out the living and bite them in order to infect them and turn them into zombies. Work together with other the zombies to overtake the station.
+
+zombie-permadeath = This time, you're dead for real.