<!-- Please read these guidelines before opening your PR: https://docs.spacestation14.io/en/getting-started/pr-guideline -->
<!-- The text between the arrows are comments - they will not be visible on your PR. -->

## About the PR
Resolve a bug where Zombies would never heal from piercing damage (guns for instance). Also the healing now scales over time in crit so we don't have bored zombie players dead for multiple minutes (up to 8 minutes before).

- Add Piercing damage (leaving it out was a mistake, did I miss any other damage types?)
- Healing scales over time the zombie is in crit, to 2x rate after 5 seconds up to 10x rate after 50 seconds. If the zombie has 200 damage (100 over crit) they will take up to 75 seconds to be up and moving again.
<!-- What does it change? What other things could this impact? -->


**Media**
<!-- 
PRs which make ingame changes (adding clothing, items, new features, etc) are required to have media attached that showcase the changes.
Small fixes/refactors are exempt.
Any media may be used in SS14 progress reports, with clear credit given.

If you're unsure whether your PR will require media, ask a maintainer.

Check the box below to confirm that you have in fact seen this (put an X in the brackets, like [X]):
-->

- [ ] I have added screenshots/videos to this PR showcasing its changes ingame, **or** this PR does not require an ingame showcase

**Changelog**
<!--
Here you can fill out a changelog that will automatically be added to the game when your PR is merged.

Only put changes that are visible and important to the player on the changelog.

Don't consider the entry type suffix (e.g. add) to be "part" of the sentence:
bad: - add: a new tool for engineers
good: - add: added a new tool for engineers

Putting a name after the :cl: symbol will change the name that shows in the changelog (otherwise it takes your GitHub username)
Like so: :cl: PJB
-->

:cl:
- tweak: Zombies heal faster while in crit, up to 10x after 50 sec. Crit will last 75 sec at most, often a lot less.
- fix: Zombies heal from Piercing (gun) damage