@tom-leys 2023-05-11
@keronshb another zombie change, this one is more minor but will lead to zombies reviving faster.