@tom-leys 2023-05-12
As per our discussion in discord @EmoGarbage404 you don't have to gib zombies to perma-kill them anymore. You just need to do an additional 50 damage or so to make killing them very likely.

A medical analyzer will show an additional 100 blunt damage when they die for real. When exactly is random to make it more interesting.

This PR overall is aimed to ensure zombie players either die permanantly or respawn quite quickly, so we don't leave players in limbo.